Hotfixes (#4)
* Fixed #1

Also fixed a weird behavior when playing a note and then moving, now the character only moves if is not playing and vice versa. This also adds another layer of control for basic_character.gd since now if the character can't move, nothing of the physics_process will happen

* Fixed #2

When changing to next scene the finished/skip/next button is disabled. Also added a disabled style box to the theme, so when the button is disabled it doesn't seem weird looking.