Hotfixes (#5)
* Fixed #1

Also fixed a weird behavior when playing a note and then moving, now the character only moves if is not playing and vice versa. This also adds another layer of control for basic_character.gd since now if the character can't move, nothing of the physics_process will happen

* Fixed #2

When changing to next scene the finished/skip/next button is disabled. Also added a disabled style box to the theme, so when the button is disabled it doesn't seem weird looking.

* Added a "null" set to object and note so they lose reference

The major problem of #1 is that after the first time the references are set, they won't reset to null, so the check would only happens successfully once. Now this may definitely fix #1

Also cleared the code removing "can_interact" logic, it isn't needed with the new approach